% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdaggregate.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate source data}
\usage{
aggregate_data(data, showprogress = FALSE)
}
\arguments{
\item{data}{A \code{cpdsourcedata} object}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
A \code{cpdaggregate} object
}
\description{
Aggregates cpdsourcedata object based on fieldtypes specified at load time.
Time period for aggregation is a calendar day.  Further options to be added at a later date.
}
