% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldtypes.R
\name{availablefieldtypes}
\alias{availablefieldtypes}
\alias{ft_timepoint}
\alias{ft_uniqueidentifier}
\alias{ft_source}
\alias{ft_categorical}
\alias{ft_number}
\alias{ft_datetime}
\alias{ft_freetext}
\alias{ft_simple}
\alias{ft_ignore}
\title{Available fieldtypes}
\usage{
ft_timepoint()

ft_uniqueidentifier()

ft_source()

ft_categorical()

ft_number()

ft_datetime()

ft_freetext()

ft_simple()

ft_ignore()
}
\description{
Specify the type of data contained in each field of your dataset using a combination of the following
}
\section{Details}{

\code{ft_timepoint} - identifies the data field which should be used as the independent time variable. There should be
one and only one of these specified.


\code{ft_uniqueidentifier} - identifies data fields which contain a (usually computer-generated) identifier for an entity, e.g. a patient.
It does not need to be unique within the dataset.


\code{ft_source} - identifies data fields which indicate that different data rows originated from distinct source systems.


\code{ft_categorical} - identifies data fields which should be treated as categorical.


\code{ft_number} - identifies data fields which contain numeric values that should be treated as continuous.


\code{ft_datetime} - identifies data fields which contain date values that should be treated as continuous.


\code{ft_freetext} - identifies data fields which contain free text values. Only presence/missingness will be evaluated.


\code{ft_simple} - identifies data fields where you only want presence/missingness to be evaluated.


\code{ft_ignore} - identifies data fields which should be ignored.  These will not be loaded.
}

\seealso{
\code{\link{fieldtypes}}
}
