% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changedetection.R
\docType{package}
\name{changedetection}
\alias{changedetection}
\alias{changedetection-package}
\title{changedetection: Automatic Change-point Detection for Large Datasets}
\description{
This is the help page for the package as a whole
}
\section{Section 1}{

Classes are S3
}

\section{Section 2}{

The functions ...
}

