% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{close_log}
\alias{close_log}
\title{Closes any active log file}
\usage{
close_log()
}
\value{
If a log file was found, the path to the log file that was closed,
  otherwise an empty string
}
\description{
Closes any active log file
}
\examples{
close_log()
}
