% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{daiquiri}
\alias{daiquiri}
\title{daiquiri: Data quality reporting for temporal datasets}
\description{
Generate reports that enable quick visual review of
temporal shifts in record-level data. Time series plots showing aggregated
values are automatically created for each column depending on its
datatype (e.g. min/max/mean values for numeric columns, no. of distinct
values for categorical columns), as well as overviews for missing values,
non-conformant values, and duplicated rows. It is designed with Electronic
Health Records in mind, but can be used for any type of record-level temporal data.
}
\details{
Classes are S3

The best place to start is the \code{\link{check_dataset}} function, and the walkthrough vignette: \href{../doc/walkthrough.html}{\code{vignette("walkthrough", package = "daiquiri")}}.
}
