% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_types.R
\name{default_field_types}
\alias{default_field_types}
\title{Create a field_types() specification with all columns one field_type}
\usage{
default_field_types(df, timepoint_field_name, default_field_type = ft_simple())
}
\arguments{
\item{df}{data frame including the column names for the template
specification}

\item{timepoint_field_name}{Name of the timepoint column in the supplied df.}

\item{default_field_type}{\code{field_type} to be used for each column. Default =
\code{\link[=ft_ignore]{ft_ignore()}}. See  \code{\link[=field_types_available]{field_types_available()}}}
}
\value{
\code{field_types()} object
}
\description{
Helper function to generate a \code{\link[=field_types]{field_types()}} specification automatically,
based on the supplied data frame. All fields (columns) in the specification
will be defined using the \code{default_field_type}, except for the \code{timepoint_field_name}.
}
\examples{
df <- data.frame(
  col1 = rep("2022-01-01", 5),
  col2 = rep(1, 5),
  col3 = 1:5,
  col4 = rnorm(5)
)

fts <-
  default_field_types(df, timepoint_field_name = "col1", default_field_type = ft_numeric())

fts
}
\seealso{
\code{\link[=field_types]{field_types()}}
}
