% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{ehrchangepoints}
\alias{ehrchangepoints}
\title{ehrchangepoints: Automatic Change-point Detection for Electronic Health Records}
\description{
This is the help page for the package as a whole
}
\section{Section 1}{

Classes are S3
}

\section{Section 2}{

The functions ...
}

