% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{export_aggregated_data}
\alias{export_aggregated_data}
\title{Export aggregated data}
\usage{
export_aggregated_data(aggregatedata, save_directory, save_filetype = "csv")
}
\arguments{
\item{aggregatedata}{A \code{aggregatedata} object}

\item{save_directory}{String. Full or relative path for save folder, use double blackslashes for nested folders and end with double backslash}

\item{save_filetype}{String. Filetype extension supported by \code{readr}, currently only csv allowed}
}
\description{
Export aggregated data to disk.  Creates a separate file for each aggregated field in dataset
}
