% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_types.R
\name{field_types}
\alias{field_types}
\title{Create field_types specification}
\usage{
field_types(..., .default_field_type = NULL)
}
\arguments{
\item{...}{names and types of fields (columns) in source data.}

\item{.default_field_type}{\code{field_type} to use for any remaining fields (columns) in source
data. Note, this means there can not be a field in the data named \code{.default_field_type}}
}
\value{
A \code{field_types} object
}
\description{
Specify the names and types of fields in the source data frame. This is
important because the data in each field will be aggregated in different
ways, depending on its \code{field_type}.  See \link{field_types_available}
}
\examples{
# specify all fields explicitly
fts <- field_types(
  PrescriptionID = ft_uniqueidentifier(),
  PrescriptionDate = ft_timepoint(),
  AdmissionDate = ft_datetime(includes_time = FALSE),
  Drug = ft_freetext(),
  Dose = ft_numeric(),
  DoseUnit = ft_categorical(),
  PatientID = ft_ignore(),
  Location = ft_categorical(aggregate_by_each_category = TRUE)
)

fts

# specify only a subset of fields
fts <- field_types(
  PrescriptionDate = ft_timepoint(),
  PatientID = ft_ignore(),
  .default_field_type = ft_simple()
)

fts
}
\seealso{
\code{\link[=field_types_available]{field_types_available()}}, \code{\link[=template_field_types]{template_field_types()}}
}
