% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldtypes.R
\name{fieldtypes}
\alias{fieldtypes}
\title{Create fieldtypes specification}
\usage{
fieldtypes(...)
}
\arguments{
\item{...}{names and types of fields (columns) in source data, in the correct order.}
}
\value{
A \code{fieldtypes} object
}
\description{
Helper function for users to specify the types of fields in the source data.  This is important both for reading
  in the data correctly, and because the data in each field will aggregated in different ways,
  depending on its fieldtype.  See \link{availablefieldtypes}
}
\examples{
fts <- fieldtypes(PatientID = ft_uniqueidentifier(),
                            TestID = ft_ignore(),
                            TestDate = ft_timepoint(),
                            TestName = ft_categorical(aggregate_by_each_category = FALSE),
                            TestResult = ft_numeric(),
                            ResultDate = ft_datetime(),
                            ResultComment = ft_freetext(),
                            Location = ft_categorical())
}
\seealso{
\code{\link{availablefieldtypes}}
}
