% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldtypes.R
\name{fieldtypes_template}
\alias{fieldtypes_template}
\title{Print a template fieldtypes specification to console}
\usage{
fieldtypes_template(df, default_fieldtype = ft_ignore())
}
\arguments{
\item{df}{data frame including the column names for the template
specification}

\item{default_fieldtype}{fieldtype to be used for each column. Default =
\code{ft_ignore()}. See  \code{\link{availablefieldtypes}}}
}
\description{
Helper function to generate template code for a fieldtypes() specification,
based on the supplied data frame. All fields (columns) in the specification
will be defined using the default_fieldtype, and the console output can be
copied and edited before being used as input to \code{\link{create_report}}()
or \code{\link{prepare_data}}().
}
\examples{
df <- data.frame(col1 = rep("2022-01-01", 5), col2 = rep(1, 5), col3 = 1:5, col4 = rnorm(5))

fieldtypes_template(df, default_fieldtype = ft_numeric())
}
\seealso{
\code{\link{fieldtypes}}
}
