% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{initialise_log}
\alias{initialise_log}
\title{Initialise a log file}
\usage{
initialise_log(log_directory)
}
\arguments{
\item{log_directory}{String containing directory to save log file}
}
\value{
Character string containing the full path to the newly-created log
file
}
\description{
Choose a directory in which to save the log file. If this is not called, no
log file is created.
}
\examples{
log_name <- initialise_log(".")

log_name
\dontshow{
close_log()
file.remove(log_name)
}
}
