% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_dataset}
\alias{load_dataset}
\title{Load source data}
\usage{
load_dataset(
  x,
  fieldtypes,
  textfile_contains_columnnames = TRUE,
  override_columnnames = FALSE,
  na = c("", "NULL"),
  showprogress = FALSE
)
}
\arguments{
\item{x}{Either a data frame or a string containing full or relative path of file containing data to load}

\item{fieldtypes}{\code{\link{fieldtypes}} object specifying names and types of fields (columns) in source data.}

\item{textfile_contains_columnnames}{If the data to be loaded is a text file, does the first row contain the column names? Default = TRUE}

\item{override_columnnames}{If FALSE, column names must exist in data frame or header row of file and must match
the names specified in fieldtypes exactly. If TRUE, column names in source will be replaced with names in fieldtypes
specification. The specification must therefore contain the columns in the correct order. Default = FALSE}

\item{na}{vector containing strings that should be interpreted as missing values, Default = \code{c("","NULL")}.}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
A \code{sourcedata} object
}
\description{
Load source data into sourcedata object
}
