% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_aggregated_data}
\alias{plot_aggregated_data}
\title{Plot data for all aggregated fields}
\usage{
plot_aggregated_data(
  cpdaggregate,
  save_plot = TRUE,
  save_directory = NULL,
  save_filetype = "png",
  showprogress = FALSE
)
}
\arguments{
\item{cpdaggregate}{A \code{cpdaggregate} object}

\item{save_plot}{Save plot to disk. Currently a bit pointless. What other options would there be?}

\item{save_directory}{String. Relative path for save folder, use double blackslashes for nested folders and end with double backslash}

\item{save_filetype}{String. Filetype extension supported by \code{ggplot2}, e.g. tif, pdf}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
A \code{cpdaggregate} object
}
\description{
Create individual plots for each data field showing any changepoints identified
}
