% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_changepoint_summary}
\alias{plot_changepoint_summary}
\title{Plot overall changepoint summary}
\usage{
plot_changepoint_summary(cpdaggregate, changepoint_methods = "all",
  save_plot = TRUE, save_directory = NULL, save_filetype = "png",
  save_filename = "Changepoint_summary", showprogress = FALSE)
}
\arguments{
\item{cpdaggregate}{A \code{cpdaggregate} object}

\item{changepoint_methods}{Vector of shortnames of changepoint methods to include or exclude from plot}

\item{save_plot}{Save plot to disk. If FALSE, opens plot in device instead}

\item{save_directory}{String. Relative path for save folder, use double blackslashes for nested folders and end with double backslash}

\item{save_filetype}{String. Filetype extension supported by \code{ggplot2}, e.g. tif, pdf}

\item{save_filename}{String. Filename for plot, excluding extension}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
A \code{cpdaggregate} object
}
\description{
Create a single plot showing changepoints identified in all data fields of aggregated object
}
