% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcedata.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare source data}
\usage{
prepare_data(
  df,
  fieldtypes,
  override_columnnames = FALSE,
  na = c("", "NA", "NULL"),
  dataset_shortdesc = NULL,
  showprogress = TRUE
)
}
\arguments{
\item{df}{A data frame}

\item{fieldtypes}{\code{\link[=fieldtypes]{fieldtypes()}} object specifying names and types
of fields (columns) in the supplied \code{df}. See also
\link{availablefieldtypes}.}

\item{override_columnnames}{If \code{FALSE}, column names in the supplied \code{df}
must match the names specified in \code{fieldtypes} exactly. If \code{TRUE}, column
names in the supplied \code{df} will be replaced with the names specified in \code{fieldtypes}.
The specification must therefore contain the columns in the correct order.
Default = \code{FALSE}}

\item{na}{vector containing strings that should be interpreted as missing
values, Default = \code{c("","NA","NULL")}.}

\item{dataset_shortdesc}{Short description of the dataset being checked. This
will appear on the report. If blank, the name of the data frame object will
be used}

\item{showprogress}{Print progress to console. Default = \code{TRUE}}
}
\value{
A \code{sourcedata} object
}
\description{
Validate a data frame against a \code{\link[=fieldtypes]{fieldtypes()}} specification, and
prepare for aggregation.
}
\examples{
# load example data into a data.frame
rawdata <- read_data(
  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

# validate and prepare the data for aggregation
sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL"),
  dataset_shortdesc = "Example data provided with package"
)
}
\seealso{
\code{\link[=fieldtypes]{fieldtypes()}}, \code{\link[=availablefieldtypes]{availablefieldtypes()}},
\code{\link[=aggregate_data]{aggregate_data()}}, \code{\link[=report_data]{report_data()}},
\code{\link[=create_report]{create_report()}}
}
