% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_data}
\alias{read_data}
\title{Read data from delimited file without doing any datatype conversion}
\usage{
read_data(
  file,
  delim = NULL,
  col_names = TRUE,
  quote = "\\"",
  trim_ws = TRUE,
  comment = "",
  skip = 0,
  n_max = Inf,
  showprogress = TRUE
)
}
\arguments{
\item{file}{A string containing path of file containing data to load, or a
URL starting \verb{http://}, \verb{file://}, etc. Compressed files with extension \code{.gz},
\code{.bz2}, \code{.xz} and \code{.zip} are supported.}

\item{delim}{Single character used to separate fields within a record. E.g.
\code{","} or \code{"\\t"}}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector of column names. If
\code{TRUE}, the first row of the input will be used as the column names, and will
not be included in the data frame. If \code{FALSE}, column names will be generated
automatically. Default = \code{TRUE}}

\item{quote}{Single character used to quote strings.}

\item{trim_ws}{Should leading and trailing whitespace be trimmed from each
field?}

\item{comment}{A string used to identify comments. Any text after the comment
characters will be silently ignored}

\item{skip}{Number of lines to skip before reading data. If \code{comment} is
supplied any commented lines are ignored after skipping}

\item{n_max}{Maximum number of lines to read.}

\item{showprogress}{Display a progress bar? Default = \code{TRUE}}
}
\value{
A data frame
}
\description{
Read rectangular data from a delimited file, with all columns read in as
character type so that the package can later detect any non-conformant
values. Operates as a restricted implementation of \code{\link[readr:read_delim]{readr::read_delim()}}.
}
\examples{
rawdata <- read_data(
  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)
}
\seealso{
\code{\link[=field_types]{field_types()}}, \code{\link[=field_types_available]{field_types_available()}},
\code{\link[=aggregate_data]{aggregate_data()}}, \code{\link[=report_data]{report_data()}},
\code{\link[=create_report]{create_report()}}
}
