% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{recalculate_changepoints}
\alias{recalculate_changepoints}
\title{Recalculate change points}
\usage{
recalculate_changepoints(data, method = "all", showprogress = FALSE)
}
\arguments{
\item{data}{An \code{aggregatedata} object}

\item{method}{String vector of changepoint methods to apply, or "all" or "none". Defaults to "all".}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
An \code{aggregatedata} object with the change points recalculated
}
\description{
Calculates change points for an aggregatedata object, replacing any existing change points previously calculated.
}
