% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_data}
\alias{report_data}
\title{Generate report from existing objects}
\usage{
report_data(
  sourcedata,
  aggregatedata,
  save_directory = ".",
  save_filename = NULL,
  format = "html",
  showprogress = TRUE
)
}
\arguments{
\item{sourcedata}{A \code{sourcedata} object returned from \code{\link{prepare_data}} function}

\item{aggregatedata}{An \code{aggregatedata} object returned from \code{\link{aggregate_data}} function}

\item{save_directory}{String specifying directory in which to save the report. Default is current directory.}

\item{save_filename}{String specifying filename for the report, excluding any file extension.
If no filename is supplied, one will be automatically generated with the format daiquiri_report_YYMMDD_HHMMSS.}

\item{format}{File format of the report. Currently only "html" is supported}

\item{showprogress}{Print progress to console. Default = TRUE}
}
\value{
A string containing the name and path of the saved report
}
\description{
Generate report from previously-created sourcedata and aggregatedata objects
}
\examples{
rawdata <- read_data(
  system.file("extdata", "abx2014.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    SourceSystem = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL"),
  showprogress = TRUE
)

aggregatedataobj <- aggregate_data(
  sourcedataobj,
  aggregation_timeunit = "day",
  showprogress = TRUE
)

report_data(
  sourcedataobj,
  aggregatedataobj,
  save_directory = ".",
  save_filename = "abx2014report",
  showprogress = TRUE
)
}
\seealso{
\code{\link{prepare_data}}, \code{\link{aggregate_data}}, \code{\link{create_report}}
}
