% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate source data}
\usage{
aggregate_data(data, changepointmethods = "all", showprogress = FALSE)
}
\arguments{
\item{data}{A \code{sourcedata} object}

\item{changepointmethods}{String vector of changepoint methods to apply, or "all" or "none". Defaults to "all".}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
A \code{aggregatedata} object
}
\description{
Aggregates sourcedata object based on fieldtypes specified at load time.
Time period for aggregation is a calendar day.  Further options to be added at a later date.
}
