% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate source data}
\usage{
aggregate_data(data, aggregation_timeunit = "day", showprogress = FALSE)
}
\arguments{
\item{data}{A \code{sourcedata} object}

\item{aggregation_timeunit}{Unit of time to aggregate over. Specify one of "day", "week", "month", "quarter", "year". The "week" option is Monday-based. Default = "day"}

\item{showprogress}{Print progress to console. Default = FALSE}
}
\value{
An \code{aggregatedata} object
}
\description{
Aggregates sourcedata object based on fieldtypes specified at load time.
Default time period for aggregation is a calendar day
}
