% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate source data}
\usage{
aggregate_data(sourcedata, aggregation_timeunit = "day", showprogress = TRUE)
}
\arguments{
\item{sourcedata}{A \code{sourcedata} object returned from \code{\link{prepare_data}} function}

\item{aggregation_timeunit}{Unit of time to aggregate over. Specify one of "day", "week", "month", "quarter", "year". The "week" option is Monday-based. Default = "day"}

\item{showprogress}{Print progress to console. Default = TRUE}
}
\value{
An \code{aggregatedata} object
}
\description{
Aggregates sourcedata object based on fieldtypes specified at load time.
Default time period for aggregation is a calendar day
}
\examples{
rawdata <- read_data(
  system.file("extdata", "example_data.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL")
)

aggregatedataobj <- aggregate_data(
  sourcedataobj,
  aggregation_timeunit = "day"
)

}
\seealso{
\code{\link{prepare_data}}, \code{\link{report_data}}
}
