% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{ehrchangepoints}
\alias{ehrchangepoints}
\title{ehrchangepoints: Temporal data quality checker for Electronic Health Records}
\description{
This package takes a generic dataset containing record-level data (i.e. one row per event, with one
column specifying the date (and time) of the event, and further columns containing any associated values for the event),
and automatically generates a report showing aggregated values for each column over time, including identifying missing values,
non-conformant values, and duplicated rows. It is designed with Electronic Health Records in mind, but can be used for any type of record-level data.
}
\details{
Classes are S3

The best place to start is the \code{\link{check_dataset}} function, and the walkthrough vignette: \href{../doc/walkthrough.html}{\code{vignette("walkthrough", package = "ehrchangepoints")}}.
}
