% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedata.R
\name{export_aggregated_data}
\alias{export_aggregated_data}
\title{Export aggregated data}
\usage{
export_aggregated_data(
  aggregatedata,
  save_directory,
  save_fileprefix = "",
  save_filetype = "csv"
)
}
\arguments{
\item{aggregatedata}{A \code{aggregatedata} object}

\item{save_directory}{String. Full or relative path for save folder}

\item{save_fileprefix}{String. Optional prefix for the exported filenames}

\item{save_filetype}{String. Filetype extension supported by \code{readr},
currently only csv allowed}
}
\description{
Export aggregated data to disk.  Creates a separate file for each aggregated
field in dataset.
}
\examples{
rawdata <- read_data(
  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL")
)

aggregatedataobj <- aggregate_data(
  sourcedataobj,
  aggregation_timeunit = "day"
)

export_aggregated_data(
  aggregatedataobj,
  save_directory = ".",
  save_fileprefix = "ex_"
)

\dontshow{
f <- list.files(".", "^ex_.*csv$")
file.remove(f)
}

}
