% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldtypes.R
\name{fieldtypes}
\alias{fieldtypes}
\title{Create fieldtypes specification}
\usage{
fieldtypes(...)
}
\arguments{
\item{...}{names and types of fields (columns) in source data.}
}
\value{
A \code{fieldtypes} object
}
\description{
Specify the names and types of fields in the source data frame. This is
important because the data in each field will be aggregated in different
ways, depending on its \code{fieldtype}.  See \link{availablefieldtypes}
}
\examples{
fts <- fieldtypes(PatientID = ft_uniqueidentifier(),
                            TestID = ft_ignore(),
                            TestDate = ft_timepoint(),
                            TestName = ft_categorical(aggregate_by_each_category = FALSE),
                            TestResult = ft_numeric(),
                            ResultDate = ft_datetime(),
                            ResultComment = ft_freetext(),
                            Location = ft_categorical())
}
\seealso{
\code{\link[=availablefieldtypes]{availablefieldtypes()}}
}
