% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldtypes.R
\name{fieldtypes}
\alias{fieldtypes}
\title{Create fieldtypes specification}
\usage{
fieldtypes(...)
}
\arguments{
\item{...}{names and types of fields (columns) in source data, in the correct order.
CHECK IF READR DEALS WITH CORRECTLY NAMED COLUMNS IN WRONG ORDER}
}
\value{
A \code{fieldtypes} object
}
\description{
Helper function for users to specify the types of fields in the source data.  This is important both for reading
  in the data correctly, and because the data in each field will be tested for changepoints in different ways,
  depending on its fieldtype.  See \link{availablefieldtypes}
}
\examples{
fts <- fieldtypes(PatientID = ft_uniqueidentifier(),
                            TestDate = ft_timepoint(),
                            TestName = ft_categorical(aggregate_by_each_category = FALSE),
                            TestResult = ft_numeric(),
                            TestComment = ft_ignore(),
                            Location = ft_categorical())
}
