% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(
  sourcedata,
  aggregatedata,
  save_directory = ".",
  save_filename = NULL,
  format = "html"
)
}
\arguments{
\item{sourcedata}{A \code{sourcedata} object returned from \code{load_dataset()} function}

\item{aggregatedata}{An \code{aggregatedata} object returned from \code{aggregate_data()} function}

\item{save_directory}{String specifying directory in which to save the report. Default is current directory.}

\item{save_filename}{String specifying filename for the report, excluding any file extension.
If no filename is supplied, one will be automatically generated with the format ehrchangepoints_report_YYMMDD_HHMMSS.}

\item{format}{File format of the report. Currently only "html" is supported}
}
\description{
Generate report from previously-created sourcedata and aggregatedata objects
}
