% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changedetection.R
\name{load_dataset}
\alias{load_dataset}
\title{Load source data}
\usage{
load_dataset(file, fieldtypes = NULL, na = NULL)
}
\arguments{
\item{file}{Full or relative path of file containing data to load}

\item{fieldtypes}{\code{\link{fieldtypes}} object specifying names and types of fields (columns) in source data.}

\item{na}{vector containing strings that should be interpreted as \code{NA}, e.g. \code{c("","NULL")}}
}
\value{
A \code{cpdsourcedata} object
}
\description{
Load source data into cpdsourcedata object
}
