% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_initialise}
\alias{log_initialise}
\title{Initialise log file}
\usage{
log_initialise(dirpath)
}
\arguments{
\item{dirpath}{String containing directory to save log file}
}
\description{
Choose a directory in which to save the log file.
If this is not called, no log file is created.
}
