% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcedata.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare source data}
\usage{
prepare_data(
  df,
  fieldtypes,
  override_columnnames = FALSE,
  na = c("", "NA", "NULL"),
  dataset_shortdesc = NULL,
  showprogress = TRUE
)
}
\arguments{
\item{df}{A data frame}

\item{fieldtypes}{\code{\link{fieldtypes}} object specifying names and types
of fields (columns) in the source data. See also
\link{availablefieldtypes}.}

\item{override_columnnames}{If FALSE, column names must exist in data frame
and must match the names specified in fieldtypes exactly. If TRUE, column
names in source will be replaced with names in fieldtypes specification.
The specification must therefore contain the columns in the correct order.
Default = FALSE}

\item{na}{vector containing strings that should be interpreted as missing
values, Default = \code{c("","NA","NULL")}.}

\item{dataset_shortdesc}{Short description of the dataset being checked. This
will appear on the report. If blank, the name of the data frame object will
be used}

\item{showprogress}{Print progress to console. Default = TRUE}
}
\value{
A \code{sourcedata} object
}
\description{
Validate a data frame against a \code{\link{fieldtypes}} specification, and
prepare for aggregation.
}
\examples{
rawdata <- read_data(
  system.file("extdata", "example_data.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL")
)
}
\seealso{
\code{\link{fieldtypes}}, \code{\link{availablefieldtypes}},
  \code{\link{aggregate_data}}, \code{\link{report_data}},
  \code{\link{create_report}}
}
