% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_data}
\alias{report_data}
\title{Generate report from existing objects}
\usage{
report_data(
  sourcedata,
  aggregatedata,
  save_directory = ".",
  save_filename = NULL,
  format = "html",
  showprogress = TRUE
)
}
\arguments{
\item{sourcedata}{A \code{sourcedata} object returned from
\code{\link[=prepare_data]{prepare_data()}} function}

\item{aggregatedata}{An \code{aggregatedata} object returned from
\code{\link[=aggregate_data]{aggregate_data()}} function}

\item{save_directory}{String specifying directory in which to save the
report. Default is current directory.}

\item{save_filename}{String specifying filename for the report, excluding any
file extension. If no filename is supplied, one will be automatically
generated with the format \code{daiquiri_report_YYMMDD_HHMMSS}.}

\item{format}{File format of the report. Currently only \code{"html"} is supported}

\item{showprogress}{Print progress to console. Default = \code{TRUE}}
}
\value{
A string containing the name and path of the saved report
}
\description{
Generate report from previously-created \code{sourcedata} and \code{aggregatedata} objects
}
\examples{
\donttest{
# load example data into a data.frame
rawdata <- read_data(
  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
  delim = ",",
  col_names = TRUE
)

# validate and prepare the data for aggregation
sourcedataobj <- prepare_data(
  rawdata,
  fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
    PrescriptionDate = ft_timepoint(),
    AdmissionDate = ft_datetime(includes_time = FALSE),
    Drug = ft_freetext(),
    Dose = ft_numeric(),
    DoseUnit = ft_categorical(),
    PatientID = ft_ignore(),
    Location = ft_categorical(aggregate_by_each_category=TRUE)),
  override_columnnames = FALSE,
  na = c("","NULL"),
  dataset_shortdesc = "Example data provided with package",
  showprogress = TRUE
)

# aggregate the data
aggregatedataobj <- aggregate_data(
  sourcedataobj,
  aggregation_timeunit = "day",
  showprogress = TRUE
)

# save a report in the current directory using the previously-created objects
report_data(
  sourcedataobj,
  aggregatedataobj,
  save_directory = ".",
  save_filename = "example_data_report",
  showprogress = TRUE
)

\dontshow{file.remove("./example_data_report.html")}
}

}
\seealso{
\code{\link[=prepare_data]{prepare_data()}}, \code{\link[=aggregate_data]{aggregate_data()}},
\code{\link[=create_report]{create_report()}}
}
